server/requirements.txt
Django==4.0.1
Pillow==9.0.0
psycopg2-binary==2.9.3
pytest-django==4.5.2